new/dev : 03 bye : 2022.08.06 12:17

+++ 03.Model2MVCShop(stu _ EL,JSTL Tag)Refactor/src/main/java/com/model2/mvc/view/cart/DeleteCartAction.java
@@ -1,8 +1,10 @@
 package com.model2.mvc.view.cart;
 
+import java.net.URLDecoder;
 import java.util.HashMap;
 import java.util.Map;
 
+import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
@@ -19,7 +21,8 @@ public class DeleteCartAction extends Action {
 	@Override
 	public String execute(HttpServletRequest request, HttpServletResponse response) throws Exception {
 		System.out.println("[DeleteCartAction execute() start...]");
-		Map<String, Object> map = new HashMap<String, Object>();
+
+		User user = (User)request.getSession(true).getAttribute("user");
 		
 		//1개 or 여러개 삭제시
 		String[] delete = request.getParameterValues("deleteCheckBox");
@@ -29,18 +32,45 @@ public class DeleteCartAction extends Action {
 			System.out.println("삭제할 상품번호 : " + deleteArray[i]);
 		}
 		
-		//삭제할 상품번호와 user_id를 map에 넣는다
-		map.put("deleteArray", deleteArray);
-		map.put("user_id", ( (User)request.getSession(true).getAttribute("user") ).getUserId() );
+		if(user == null || user.getUserId().equals("non-member")) {
+			// 비회원이라면
+			String allInfo = "";
+			
+			Cookie[] cookies = request.getCookies();
+			if(cookies != null && cookies.length > 0) {
+				for (int i = 0; i < cookies.length; i++) {
+					if(cookies[i].getName().equals("prodInfoCookie")) {
+						allInfo = URLDecoder.decode(cookies[i].getValue());
+					}
+				}
+			}
+			
+			String testStr = "10001:2,10003:5,10007:12";
+			int testIndex = testStr.indexOf("10003");
+			System.out.println("return된 index : " + testIndex);
+			
+			// , 전까지 자른다 , 가 없다면 index부터 끝까지 자른다
+			
+			//request.setAttribute("list", map.get("list"));
+			//count : 게시물 수, listCart.jsp에서 count>0일때 for문으로 list출력
+			//request.setAttribute("count", map.get("count"));
+		}else {
+			//회원이라면
+			Map<String, Object> map = new HashMap<String, Object>();
+			
+			//삭제할 상품번호와 user_id를 map에 넣는다
+			map.put("deleteArray", deleteArray);
+			map.put("user_id", ( (User)request.getSession(true).getAttribute("user") ).getUserId() );
 
-		//장바구니에서 상품을 삭제하고 삭제한 list를 가져온다
-		CartService service = new CartServiceImpl();
-		service.deleteCart(map);
-		map = service.getCartList( ( (User)request.getSession(true).getAttribute("user") ).getUserId() );
-		
-		request.setAttribute("list", map.get("list"));
-		//count : 게시물 수, listCart.jsp에서 count>0일때 for문으로 list출력
-		request.setAttribute("count", map.get("count"));
+			//장바구니에서 상품을 삭제하고 삭제한 list를 가져온다
+			CartService service = new CartServiceImpl();
+			service.deleteCart(map);
+			map = service.getCartList( ( (User)request.getSession(true).getAttribute("user") ).getUserId() );
+			
+			request.setAttribute("list", map.get("list"));
+			//count : 게시물 수, listCart.jsp에서 count>0일때 for문으로 list출력
+			request.setAttribute("count", map.get("count"));
+		}
 		
 		System.out.println("[DeleteCartAction execute() end...]");
 		return "forward:/cart/listCart.jsp";
